+++ packages/desktop-client/src/i18n.ts
@@ -28,29 +28,3 @@ i18n
       transSupportBasicHtmlNodes: false,
     },
   });
-
-const interleaveArrays = (
-  ...arrays: (React.ReactNode | string)[][]
-) =>
-  Array.from(
-    {
-      length: Math.max(...arrays.map(array => array.length)),
-    },
-    (_, i) => arrays.map(array => array[i]),
-  ).flat();
-
-export const i18nObjectList = (values: (React.ReactNode | string)[], lng: string, opt = {}) => {
-  const formatter = new Intl.ListFormat(lng, {
-    style: 'long',
-    type: 'conjunction',
-    ...opt,
-  });
-
-  const placeholders = Array.from(
-    { length: values.length },
-    (_, i) => `<${i}>`,
-  );
-  const formatted = formatter.format(placeholders);
-  const parts = formatted.split(/<\d+>/g);
-  return interleaveArrays(parts, values);
-};
